refactor: use analytics aliases
